api/requirements.txt
flask==2.0.3
gunicorn==20.1.0
pytest==7.1.1
redis==4.1.4
werkzeug==2.0.3
opencv-python>=4.1.1
flask_httpauth==4.7
flask_login==0.6.2
Pillow==9.2.0